+++ src/nice_elapsed/mod.rs
@@ -2,16 +2,15 @@
 # Dactyl: "Nice" Elapsed
 */
 
+pub(super) mod clock;
+
 use crate::{
 	NiceU16,
 	traits::SaturatingFrom,
 };
 use std::{
 	fmt,
-	hash::{
-		Hash,
-		Hasher,
-	},
+	hash,
 	ops::Deref,
 	time::{
 		Duration,
@@ -23,20 +22,6 @@ use std::{
 
 /// # Array Size.
 const SIZE: usize = 52;
-
-/// # Helper: `AsRef` and `Borrow`.
-macro_rules! as_ref_borrow_cast {
-	($($cast:ident $ty:ty),+ $(,)?) => ($(
-		impl AsRef<$ty> for NiceElapsed {
-			#[inline]
-			fn as_ref(&self) -> &$ty { self.$cast() }
-		}
-		impl ::std::borrow::Borrow<$ty> for NiceElapsed {
-			#[inline]
-			fn borrow(&self) -> &$ty { self.$cast() }
-		}
-	)+);
-}
 
 /// # Helper: Generate Impl
 macro_rules! elapsed_from {
@@ -69,6 +54,8 @@ macro_rules! elapsed_from {
 /// `From<Duration>` or `From<Instant>`, in which case milliseconds (to two
 /// decimal places) will be included, unless zero.
 ///
+/// For a more clock-like output, see [`NiceClock`](crate::NiceClock).
+///
 /// ## Examples
 ///
 /// ```
@@ -91,7 +78,15 @@ impl AsRef<[u8]> for NiceElapsed {
 	fn as_ref(&self) -> &[u8] { self.as_bytes() }
 }
 
-as_ref_borrow_cast!(as_str str);
+impl AsRef<str> for NiceElapsed {
+	#[inline]
+	fn as_ref(&self) -> &str { self.as_str() }
+}
+
+impl ::std::borrow::Borrow<str> for NiceElapsed {
+	#[inline]
+	fn borrow(&self) -> &str { self.as_str() }
+}
 
 impl Default for NiceElapsed {
 	#[inline]
@@ -105,6 +100,7 @@ impl Default for NiceElapsed {
 
 impl Deref for NiceElapsed {
 	type Target = [u8];
+
 	#[inline]
 	fn deref(&self) -> &Self::Target { self.as_bytes() }
 }
@@ -167,9 +163,9 @@ impl From<u32> for NiceElapsed {
 // These all work the same way.
 elapsed_from!(usize, u64, u128);
 
-impl Hash for NiceElapsed {
+impl hash::Hash for NiceElapsed {
 	#[inline]
-	fn hash<H: Hasher>(&self, state: &mut H) { state.write(self.as_bytes()); }
+	fn hash<H: hash::Hasher>(&self, state: &mut H) { state.write(self.as_bytes()); }
 }
 
 impl PartialEq for NiceElapsed {